Add condition to message deletion

+++ src/helpers/messageHelper.ts
@@ -1,8 +1,13 @@
 import { ButtonInteraction, CommandInteraction } from "discord.js";
 
+let justDeletedId;
+
 const deleteReply = (interaction: ButtonInteraction | CommandInteraction, timeBeforeDeletion: number) => {
   setTimeout(async () => {
-    await interaction.deleteReply();
+    if (justDeletedId !== interaction.id) {
+      await interaction.deleteReply();
+      justDeletedId = interaction.id;
+    }
   }, timeBeforeDeletion);
 };
